requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.8
click==8.0.3
Deprecated==1.2.13
Flask==2.0.2
Flask-Caching==1.10.1
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
redis==4.0.2
requests==2.26.0
urllib3==1.26.7
Werkzeug==2.0.2
wrapt==1.13.3
plaid_python==8.2.1
python-dotenv==0.15.0